+++ build-tools/packages/build-tools/src/common/fluidRepo.ts
@@ -183,6 +183,11 @@ export interface IFluidRepoPackage {
 	 */
 	ignoredDirs?: string[];
 
+	/**
+	 * The interdependencyRange controls the type of semver range to use between packages in the same release group. This
+	 * setting controls the default range that will be used when updating the version of a release group. The default can
+	 * be overridden using the `--interdependencyRange` flag in the `flub bump` command.
+	 */
 	defaultInterdependencyRange: InterdependencyRange;
 }
